+++ src/model/mod.rs
@@ -1,3 +1,4 @@
+pub mod backup;
 pub mod event;
 pub mod pomodoro;
 pub mod tag;